+++ changelog.md
@@ -1,5 +1,12 @@
 # Base Change Log
 
+## [4.1.0] - 2018-12-10
+### Base V4.1.0
+- Removed gulp and replaced with webpack
+- Updated scripts in package.json
+- Minor documentation fixes
+- Version bump for index.scss/less and package.json
+
 ## [4.0.10] - 2018-05-17
 ### Base V4.0.10
 - Added Postinstall script to fix normalize.css import issue